Add Cast::dynamic_cast()
This does all type-checking at runtime and is required because it's not
always statically known at compile-time whether an object implements an
interface or not.

Also add notes about this limitation to Cast::downcast() and Cast::upcast().

And change Cast::is() to only perform runtime checking. It was always
doing runtime checking anyway, but in addition did compile-time checking
which is more restricted. If compile-time checking is wanted in
addition, Downcast::can_downcast() can be used for the old behaviour.

https://github.com/gtk-rs/glib/issues/205