Fix/improve the translation from GLib paths to Rust paths and back
First of all, add comments in the Windows and UNIX specific cases about
what a GLib path is, and what a Rust path is.

On UNIX, paths are in the local encoding in GLib and in Rust and must
never contain NUL bytes except for the final NUL terminator. As such we
can safely panic if this is ever not true, there's nothing else we can
do in this case and the previous code already implicitly did that by
unwrapping.

On Windows, GLib paths are always UTF-8 and Rust paths are WTF-8,
which is a superset of UTF-8 and not every WTF-8 string can be converted
back to UTF-8. As such, conversion from Rust to GLib can fail and we
can't come up with any UTF-8 string that would represent the same path
and have to panic (and did so implicitly by unwrapping before already).
Conversion from GLib to Rust can't possibly fail, unless there is a bug
in the C library and the string is not valid UTF-8. In this case,
instead of replacing the invalid characters with the replacement
character, the best we can do is to panic. By using the replacement
characters, in the worst case a different (and existing!) path could be
used, resulting in undefined behaviour and potential data loss.