Remove Default impl for SourceId
While 0 is never a valid SourceId, it is also not valid to use it with
any functions that work on SourceIds. It causes assertions at runtime.
The pattern of having a "non-existing" SourceId is better suited by
using an Option, and once NonZero is stabilized this can even be done
without introducing any memory overhead.