Provide manual bindings for GString.
C API follows chaining pattern which is replicated in Rust version as
well. It is manually written instead of being auto-generated to avoid
unnecessary copies (generator doesn't know that the same thing is being
returned back as was provided to a function as a mutable argument).

GString is intended to represent textual content, but there is no
guarantee that data it contains will valid UTF-8. For convenience two
conversion methods are provided to_str and to_string_lossy, based
on interface in Rust CStr.