Add new Object::connect_unsafe(), Object::connect_notify_unsafe() and Closure::new_unsafe()
These allow to use non-Send/non-Sync/non-'static closures and it's the
job of the caller to ensure that the closures are only used in a safe
way.

Also clean up the calculation of the size of the allocated GClosure
object, and as a side-effect allocate 4 bytes less on 32 bit
architectures.