RFC: Add Array<T>
It's not clear to me how much magic should be done when manipulating a
GArray. I avoided glib pointer conversion for inner elements for now
(see GPtrArray for that). I wonder if there is a trait to say that a
type can be safely Copy from its pointer (which is the GArray append
API)

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>