Remove uninitialized() function for Boxed types
These would currently allocate a Box<_> for the FFI struct and would be
used for boxed type out function parameters, where the caller is
supposed to allocate the memory.

The problem with these is that if the FFI struct requires a function to
free memory stored inside it, we have no way of freeing it currently and
would leak the memory.

As such remove the support for this until it is actually needed, and
then we can figure out a solution based on the concrete case.

Fixes https://github.com/gtk-rs/glib/issues/469