Add checking for properties in Object::new()
We didn't check if a property of that name exists or if the given value
is valid for that property. While GObject is doing that itself already,
its type checks are less strict than ours, which caused inconsistencies
with set_property() for example and is simply unexpected.

Fixes https://github.com/gtk-rs/glib/issues/641