Add DynamicVariantType
Currently Variant API is limited to StaticVariantType, that can only be
implemented by types with static+fixed signatures. This will not work
for generic container types, where the signature is dependent on the child
type.

This patch adds a new trait that allows the type signature to be created
dynamically. Blanket implementation is provided for common container
types in Rust.