Add a ObjectSubclass version of glib_wrapper!
This makes it easier to define a wrapper for a type defined in Rust
using `ObjectSubclass`. It also updates the `glib::subclass`
documentation to use this.

I'm not all that fond of the `glib_wrapper!` macro, and would be happy
to see something better, but this seems like a good way to improve
subclassing ergonomics for now.

There's certainly more room for improvement in both the API and
documentation for subclassing.