Derive PartialOrd, PartialEq, Eq, and Debug in glib_object_wrapper!
I believe the derived `Ord` and `Eq` implementations should behave
correctly based on how the wrapper `ObjectRef` type sorts.

The derived `Debug` changes this to printing as a tuple struct, rather
than a struct with an `inner` member. If this is undesired, the struct
should just be defined to actually have an `inner` member.